requirements.txt
python-telegram-bot[job-queue]==21.6
fastapi==0.121.3
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
asyncpg==0.29.0
httpx==0.27.2
requests==2.32.3
pydantic==2.9.2
pytz==2024.1
